fixed minor double to float warnings in getJetColor helper.

+++ src/util/utils.cc
@@ -189,17 +189,17 @@ Eigen::Vector4f getJetColor(double v, double vmin, double vmax)
    dv = vmax - vmin;
 
    if (v < (vmin + 0.25 * dv)) {
-      c[0] = 0;
-      c[1] = 4 * (v - vmin) / dv;
+      c[0] = 0.;
+      c[1] = static_cast<float>(4. * (v - vmin) / dv);
    } else if (v < (vmin + 0.5 * dv)) {
-      c[0] = 0;
-      c[2] = 1 + 4 * (vmin + 0.25 * dv - v) / dv;
+      c[0] = 0.;
+      c[2] = static_cast<float>(1. + 4. * (vmin + 0.25 * dv - v) / dv);
    } else if (v < (vmin + 0.75 * dv)) {
-      c[0] = 4 * (v - vmin - 0.5 * dv) / dv;
-      c[2] = 0;
+      c[0] = static_cast<float>(4. * (v - vmin - 0.5 * dv) / dv);
+      c[2] = 0.;
    } else {
-      c[1] = 1 + 4 * (vmin + 0.75 * dv - v) / dv;
-      c[2] = 0;
+      c[1] = static_cast<float>(1. + 4. * (vmin + 0.75 * dv - v) / dv);
+      c[2] = 0.;
    }
 
    return c;
